codebase/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg==3.1.18
django-allauth==0.62.1 
django-crispy-forms==2.1
django-filter==24.2
django-ckeditor==6.7.1
django-countries==7.6.1
django-leaflet==0.29.1
django-extensions==3.2.3
django-widget-tweaks==1.5.0
pandas==2.2.2
geopy==2.4.1
shapely==2.0.4
folium==0.16.0
django-cleanup==8.1.0
Pillow==10.3.0
django-sass-processor==1.4
drf-yasg==1.21.7